+++ frontend/docs/jinhokim98.md
@@ -135,7 +135,7 @@
       - [x] 이미지가 있다면 이미지 height와 gap 처리
       - [x] 대댓글 버튼 클릭 시 대댓글 달기 모드로 변환
       - [x] isMine 추가로 삭제 수정버튼 보이도록 설정
-      - [ ] 글 수정 폼 추가
+      - [x] 글 수정 폼 추가 (본인의 글이 아닐 경우 비정상적인 접근처리)
       
       - api 호출
         - [x] 글 상세 조회
