+++ frontend/src/components/community/detail/content/Content.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import WriteInfo from '@components/community/common/WriteInfo';
 import { ArticleDetailType } from '@components/community/CommunityType';
 import Buttons from '@components/community/common/Buttons';
+import { useNavigate } from 'react-router-dom';
+import ROUTE from '@constants/route';
 import * as C from './Content.style';
 
 interface ContentProps {
@@ -10,7 +12,10 @@ interface ContentProps {
 }
 
 function Content({ id, data }: ContentProps) {
-  const updateArticle = (articleId: number) => {};
+  const navigate = useNavigate();
+  const updateArticle = () => {
+    navigate(`${ROUTE.COMMUNITY.BASE}/${ROUTE.COMMUNITY.UPDATE}`, { state: data });
+  };
   const deleteArticle = (articleId: number) => {};
 
   return (
@@ -22,7 +27,7 @@ function Content({ id, data }: ContentProps) {
       ))}
       <C.Content>{data.content}</C.Content>
       <Buttons
-        normal={{ label: '수정', onClick: () => updateArticle(id), show: data.isMine }}
+        normal={{ label: '수정', onClick: updateArticle, show: data.isMine }}
         dangerous={{ label: '삭제', onClick: () => deleteArticle(id), show: data.isMine }}
       />
     </>
