+++ frontend/src/components/community/index.ts
@@ -1,3 +1,4 @@
 export { default as CommunityList } from './list/CommunityList';
 export { default as CommunityWrite } from './write/CommunityWrite';
 export { default as ArticleDetail } from './detail/ArticleDetail';
+export { default as CommunityUpdate } from './update/CommunityUpdate';
