+++ frontend/src/components/community/write/CommunityWrite.style.tsx
@@ -35,97 +35,6 @@ export const InputTitle = styled.input`
   }
 `;
 
-export const Textarea = styled.textarea`
-  width: 100%;
-  min-height: 32.5rem;
-  padding: 1rem 2rem;
-  border-radius: 0.8rem;
-  border: none;
-
-  color: ${(props) => props.theme.colors.white};
-
-  background: ${(props) => props.theme.colors.grey001};
-  ${(props) => props.theme.typography[DeviceProvider()].body};
-  overflow-y: hidden;
-  outline: none;
-  resize: none;
-
-  &::placeholder {
-    color: ${(props) => props.theme.colors.grey002};
-  }
-`;
-
-export const ImageContainer = styled.aside`
-  margin-bottom: 4rem;
-`;
-
-export const Label = styled.div`
-  margin-bottom: 1rem;
-  color: ${(props) => props.theme.colors.white};
-  ${(props) => props.theme.typography[DeviceProvider()].body};
-`;
-
-export const ImageUploads = styled.div`
-  display: flex;
-  width: 100%;
-  overflow-x: scroll;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-export const ImageView = styled.img`
-  width: 6rem;
-  height: 6rem;
-  flex-shrink: 0;
-  margin-right: 1rem;
-
-  border-radius: 1rem;
-  object-fit: cover;
-
-  ${(props) => props.theme.media.tablet`
-    width: 8rem;
-    height: 8rem;
-  `};
-
-  ${(props) => props.theme.media.desktop`
-    width: 10rem;
-    height: 10rem;
-  `};
-
-  ${(props) => props.theme.media.wide`
-    width: 10rem;
-    height: 10rem;
-  `};
-`;
-
-export const ImageUploadButton = styled.div<{ hide: boolean }>`
-  display: ${(props) => (props.hide ? 'none' : 'flex')};
-  justify-content: center;
-  align-items: center;
-  width: 6rem;
-  height: 6rem;
-  flex-shrink: 0;
-
-  border-radius: 1rem;
-  background: ${(props) => props.theme.colors.grey001};
-
-  ${(props) => props.theme.media.tablet`
-    width: 8rem;
-    height: 8rem;
-  `};
-
-  ${(props) => props.theme.media.desktop`
-    width: 10rem;
-    height: 10rem;
-  `};
-
-  ${(props) => props.theme.media.wide`
-    width: 10rem;
-    height: 10rem;
-  `};
-`;
-
 export const UploadButton = styled.button`
   display: flex;
   justify-content: center;
