+++ frontend/src/components/community/write/EachImage.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import * as EI from './CommunityWrite.style';
+import ImageView from './EachImage.style';
 
 interface EachImageProps {
   image: string;
 }
 
 function EachImage({ image }: EachImageProps) {
-  return <EI.ImageView src={image} alt="image" />;
+  return <ImageView src={image} alt="image" />;
 }
 
 export default EachImage;
