+++ frontend/src/constants/route.ts
@@ -5,6 +5,7 @@ const ROUTE = Object.freeze({
     BASE: '/community',
     LIST: '',
     WRITE: 'write',
+    UPDATE: 'update',
     DETAIL: ':id',
     NOT_FOUND: '404-not-found',
   },
